refactor(prisma): rename `type.ts` into `utils.ts` - Move area guard functions from `helper.ts` into `utils.ts`
