+++ prisma/helper.ts
@@ -1,15 +1,5 @@
 import { createReadStream } from 'fs';
 import * as Papa from 'papaparse';
-import { Areas, District, Regency, Village } from './types';
-
-export const isRegency = (area: Areas): area is Regency =>
-  'province_code' in area;
-
-export const isDistrict = (area: Areas): area is District =>
-  'regency_code' in area;
-
-export const isVillage = (area: Areas): area is Village =>
-  'district_code' in area;
 
 /**
  * Parse CSV from local file asynchronously.
