+++ prisma/utils.ts
@@ -34,3 +34,12 @@ export type AreaByCollection<Collection> = Collection extends 'provinces'
   : Collection extends 'villages'
   ? Village
   : never;
+
+export const isRegency = (area: Areas): area is Regency =>
+  'province_code' in area;
+
+export const isDistrict = (area: Areas): area is District =>
+  'regency_code' in area;
+
+export const isVillage = (area: Areas): area is Village =>
+  'district_code' in area;
